chore: remove obsolete comments
