+++ controllers/lighthouse.js
